Creating PluginDocumentation.tsx component;
